# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/users/register": {
      "post": {
        "operationId": "UsersController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "409": {
            "description": "Email already registered"
          }
        },
        "summary": "Register a new user (teacher or parent)",
        "tags": [
          "Users"
        ]
      }
    },
    "/users/login": {
      "post": {
        "operationId": "UsersController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Credentials valid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Invalid email or password"
          }
        },
        "summary": "Validate credentials and return the user profile with a signed JWT",
        "tags": [
          "Users"
        ]
      }
    },
...
  },
  "info": {
    "title": "SessionGrid API",
    "description": "Global Class Offering Booking System — teachers create courses and sessions, parents book them with full timezone support and race-condition-proof concurrency.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "x-user-id": {
        "type": "apiKey",
        "in": "header",
        "name": "x-user-id"
      },
      "x-user-role": {
        "type": "apiKey",
        "in": "header",
        "name": "x-user-role"
      },
      "x-user-timezone": {
        "type": "apiKey",
        "in": "header",
        "name": "x-user-timezone"
      }
    },
    "schemas": {
      "RegisterUserDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "alice@example.com"
          },
          "firstName": {
            "type": "string",
            "example": "Alice"
          },
          "lastName": {
            "type": "string",
            "example": "Smith"
          },
          "password": {
            "type": "string",
            "example": "secret123",
            "minLength": 8
          },
          "role": {
            "type": "string",
            "enum": [
              "teacher",
              "parent"
            ],
            "example": "teacher"
          },
          "timezone": {
            "type": "string",
            "example": "Asia/Kolkata",
            "description": "IANA timezone string used for all time conversions"
          }
        },
        "required": [
          "email",
          "firstName",
          "lastName",
          "password",
          "role",
          "timezone"
        ]
      },
      "UserResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "3f9e0c8d-6ce8-492f-99c9-76fb3b0c62aa"
          },
          "email": {
            "type": "string",
            "example": "alice@example.com"
          },
          "firstName": {
            "type": "string",
            "example": "Alice"
          },
          "lastName": {
            "type": "string",
            "example": "Smith"
          },
          "role": {
            "type": "string",
            "enum": [
              "teacher",
              "parent"
            ],
            "example": "teacher"
          },
          "timezone": {
            "type": "string",
            "example": "Asia/Kolkata"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2024-09-01T09:00:00.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2024-09-01T09:00:00.000Z"
          }
        },
        "required": [
          "id",
          "email",
          "firstName",
          "lastName",
          "role",
          "timezone",
          "createdAt",
          "updatedAt"
        ]
      },
      "LoginUserDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "alice@example.com"
          },
          "password": {
            "type": "string",
            "example": "secret123"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "LoginResponseDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiI...",
            "description": "HS256-signed JWT. Pass as Authorization: Bearer <token> on protected routes."
          },
          "user": {
            "$ref": "#/components/schemas/UserResponseDto"
          }
        },
        "required": [
          "accessToken",
          "user"
        ]
      },
...
    }
  }
}